Allow to provide GCP service account programmatically.
This allows to avoid relying on environment variables and provide data
programmatically in environments where changing environment variables is
not feasible.

Also add --credentials_file argument which allows to provide a path to
GCP service account via argument. This is mostly to be able to easily
test this new change.